+++ exercise3/local_src/game-1.0/test.c
@@ -19,7 +19,7 @@ void test_spawn_fruit(){
 	//Spawn three fruits
 	printf("Spawning three fruits and printing the board\n");
 	//BOARD = {0}{0};
-	memset(BOARD, 0, sizeof(int)*BOARD_SIZE*BOARD_SIZE);
+	memset(BOARD, 0, sizeof(uint16_t)*BOARD_SIZE*BOARD_SIZE);
 	spawn_fruit();
 	spawn_fruit();
 	spawn_fruit();
@@ -31,8 +31,8 @@ void test_spawn_fruit(){
 void test_initialize_snake(){
 	printf("initialize Snake and printing board\n");
 	//BOARD = {0}{0};
-	memset(BOARD, 0, sizeof(int)*BOARD_SIZE*BOARD_SIZE);
-	int headPos[] = {15, 15};
+	memset(BOARD, 0, sizeof(uint16_t)*BOARD_SIZE*BOARD_SIZE);
+	int8_t headPos[] = {15, 15};
 	initialize_snake(headPos);
 	print_board();
 }
@@ -40,9 +40,9 @@ void test_initialize_snake(){
 
 void test_reorder_snake(){
 	printf("First: initialize a snake");
-	int head[2] = {15, 15};
+	int8_t head[2] = {15, 15};
 	//BOARD = {0}{0};
-	memset(BOARD, 0, sizeof(int)*BOARD_SIZE*BOARD_SIZE);
+	memset(BOARD, 0, sizeof(uint16_t)*BOARD_SIZE*BOARD_SIZE);
 	initialize_snake(head);
 	printf("Second: move snake DOWNL");
 	snake_movement(head, DOWNL);
@@ -54,9 +54,9 @@ void test_reorder_snake(){
 
 void test_snake_movement(){
 	// DOWNL
-	int8_t i;
-	memset(BOARD, 0, sizeof(int)*BOARD_SIZE*BOARD_SIZE);
-	int headPos[2] = {15, 15};
+	uint8_t i;
+	memset(BOARD, 0, sizeof(uint16_t)*BOARD_SIZE*BOARD_SIZE);
+	int8_t headPos[2] = {15, 15};
 	printf("Old headPos:\n");
 	for(i = 0; i<2; i++){
 		printf("%d:", i);
@@ -69,7 +69,7 @@ void test_snake_movement(){
 		printf("%d\n" , headPos[i]);
 	}
 	// UPL
-	memset(BOARD, 0, sizeof(int)*BOARD_SIZE*BOARD_SIZE);
+	memset(BOARD, 0, sizeof(uint16_t)*BOARD_SIZE*BOARD_SIZE);
 	headPos[0] = 15;
 	headPos[1] =15;
 	printf("Old headPos:\n");
@@ -84,7 +84,7 @@ void test_snake_movement(){
 		printf("%d\n" , headPos[i]);
 	}
 	// RIGHT
-	memset(BOARD, 0, sizeof(int)*BOARD_SIZE*BOARD_SIZE);
+	memset(BOARD, 0, sizeof(uint16_t)*BOARD_SIZE*BOARD_SIZE);
 	headPos[0] = 15;
 	headPos[1] =15;
 	printf("Old headPos:\n");
@@ -100,7 +100,7 @@ void test_snake_movement(){
 	}
 	//LEFT 
 	
-	memset(BOARD, 0, sizeof(int)*BOARD_SIZE*BOARD_SIZE);
+	memset(BOARD, 0, sizeof(uint16_t)*BOARD_SIZE*BOARD_SIZE);
 	headPos[0] = 15;
 	headPos[1] =15;
 	printf("Old headPos:\n");
